Only print stdout buf of runner if it exists
This fixes a crash that happens when quitting fzz with Ctrl-D or return
right after starting it. The problem was that quitting would result in
runner trying to send the buffered stdout of the last command to the
printer, but since no command has been run, the stdout buf does not yet
exist.

This checks if it exists and only prints then, otherwise it silently
returns.