Validate placeholder when starting up
Since we can set a placeholder with the environment variable
FZZ_PLACEHOLDER instead of using the default placeholder we now validate
this custom placeholder.

The plan is to allow users to specify initial input with the
placeholder, like this:

  fzz ag {{foobar}} *.go

In order to to that the number of characters in the placeholder has to
be valid.