Fixes printing of too many too wide lines
The printer would cut off the `newline` character at the end of the last
line in the `Print()` method, if it was too wide.

The `printLine()` method however would re-add the newline character,
even if it was the last line, if the line was too wide. It would cut the
line down to `maxCol-1` and then add the newline.

This caused the whole output to be shifted one line upwards in the
terminal, which hid the prompt.